Rename perf to hbench
The main reason for the rename is that perf (and performance measuring
tool description) suggests that perf is a generic tool for measuring
performance of other applications. But it is actually a set of
benchmarks for measuring internal parts of HelenOS (libc, IPC, etc.).

Furthermore, perf is a well-known program in GNU/Linux world with
completely different purpose and there really is no reason why to
confuse the users.

The new name hopefully better suggests that the tool is a set of
benchmarks for HelenOS (hence the name (H)elenOS (bench)marks).
